Refactor `ContactRepository` to use interface
